+++ NEWS.md
@@ -3,3 +3,8 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Added a new vignette `vignette("bootstrapping_many_curves")` to demonstrate how to run the bootstrap procedure on multiple curves at once
 * Updated README to reflect this
+
+# rTPC 1.0.2 - 03/12/2021
+
+* Added the Lobry–Rosso–Flandrois model to the package
+* Updated all other functions to include the lrf_1991() model